requirements.txt
# Telegram Bot
aiogram>=3.13.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0

# Google Sheets
gspread>=6.1.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0

# Excel
openpyxl>=3.1.0

# HTTP client
httpx>=0.27.0

# Forms (FastAPI file uploads)
python-multipart>=0.0.22

# Tests
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Utils
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
